Logs implementation
Logs implemented in CityController
