--- src/main/java/com/brickncode/backendarchetype/BackendArchetypeApplication.java
@@ -1,14 +1,0 @@
-package com.brickncode.backendarchetype;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-
-@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
-public class BackendArchetypeApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackendArchetypeApplication.class, args);
-	}
-
-}
