--- src/main/java/com/brickncode/backendarchetype/common/URLConstant.java
@@ -1,11 +1,0 @@
-package com.brickncode.backendarchetype.common;
-
-public interface URLConstant {
-	String CONTROLLER_PREFIX = "/world";
-	String GET_ALL = "/get-all";
-	String GET_ONE = "/get-one";
-	String POST = "/post";
-	String PUT = "/put";
-	String DELETE = "/delete";
-	String PATH_ID_VARIABLE = "/{id}";
-}
