--- src/main/java/com/brickncode/backendarchetype/data/dto/CountryClass.java
@@ -1,9 +1,0 @@
-package com.brickncode.backendarchetype.data.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-public class CountryClass {
-
-	@Getter @Setter private String code;
-}
