--- src/main/java/com/brickncode/backendarchetype/data/entity/enums/Continent.java
@@ -1,9 +1,0 @@
-package com.brickncode.backendarchetype.data.entity.enums;
-
-public enum Continent {
-	ASIA("Asia"), EUROPE("Europe"), NORTH_AMERICA("North America"),
-	AFRICA("Africa"), OCEANIA("Oceania"), ANTARCTICA("Antarctica"),
-	SOUTH_AMERICA("South America");
-
-	private Continent(String value) { }
-}
