--- src/main/java/com/brickncode/backendarchetype/repository/impl/SampleRepository.java
@@ -1,4 +1,0 @@
-package com.brickncode.backendarchetype.repository.impl;
-
-public class SampleRepository {
-}
