+++ src/main/java/com/bytelius/backendarchetype/city/data/dto/output/CityOutput.java
@@ -1,0 +1,18 @@
+package com.bytelius.backendarchetype.city.data.dto.output;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class CityOutput {
+
+	@Getter @Setter private Integer id;
+
+	@Getter @Setter private String name;
+
+	public CityOutput() { }
+
+	public CityOutput(Integer id, String name) {
+		this.id = id;
+		this.name = name;
+	}
+}
