+++ src/main/java/com/bytelius/backendarchetype/city/repository/impl/CityRepository.java
@@ -1,0 +1,5 @@
+package com.bytelius.backendarchetype.city.repository.impl;
+
+public class CityRepository {
+
+}
