+++ src/main/java/com/bytelius/backendarchetype/country/data/dto/CountryClass.java
@@ -1,0 +1,9 @@
+package com.bytelius.backendarchetype.country.data.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+public class CountryClass {
+
+	@Getter @Setter private String code;
+}
