+++ src/main/java/com/bytelius/backendarchetype/country/data/entity/enums/Continent.java
@@ -1,0 +1,9 @@
+package com.bytelius.backendarchetype.country.data.entity.enums;
+
+public enum Continent {
+	ASIA("Asia"), EUROPE("Europe"), NORTH_AMERICA("North America"),
+	AFRICA("Africa"), OCEANIA("Oceania"), ANTARCTICA("Antarctica"),
+	SOUTH_AMERICA("South America");
+
+	private Continent(String value) { }
+}
