+++ src/main/java/com/bytelius/backendarchetype/shared/GenericResponse.java
@@ -1,0 +1,22 @@
+package com.bytelius.backendarchetype.shared;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.util.MultiValueMap;
+
+public class GenericResponse extends ResponseEntity {
+
+	@Getter @Setter private String message;
+
+	public GenericResponse(HttpStatus status, String message) {
+		super(status);
+		this.message = message;
+	}
+
+	public GenericResponse(HttpStatus status, String message, Object body) {
+		super(body, status);
+		this.message = message;
+	}
+}
