+++ src/main/java/com/bytelius/backendarchetype/shared/URLConstant.java
@@ -1,0 +1,11 @@
+package com.bytelius.backendarchetype.shared;
+
+public interface URLConstant {
+	String CONTROLLER_PREFIX = "/world";
+	String GET_ALL = "/get-all";
+	String GET_ONE = "/get-one";
+	String POST = "/post";
+	String PUT = "/put";
+	String DELETE = "/delete";
+	String PATH_ID_VARIABLE = "/{id}";
+}
