+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 #Database connexion
-spring.datasource.url=jdbc:mysql://localhost:3306/world
+spring.datasource.url=jdbc:mysql://localhost:3306/world2
 spring.datasource.username=root
 spring.datasource.password=root
 #spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
@@ -15,5 +15,5 @@ spring.liquibase.change-log=classpath:/liquibase/db.changelog-master-integration
 server.port=8090
 
 #Configure logs path
-logging.file.path=./logs
+logging.file.path=./log
 logging.file.name=${logging.file.path}/application.log
