+++ src/main/resources/liquibase/db.changelog-master-integration.xml
@@ -4,8 +4,8 @@
     xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.0.xsd
     http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd">
 	
-	<!--<include file="changelog/db.changelog-property-definition.xml" relativeToChangelogFile="true"/>
+	<!--<include file="changelog/db.changelog-property-definition.xml" relativeToChangelogFile="true"/>-->
 	<include file="changelog/ddl/db.changelog-ddl-v1.0.0.xml" relativeToChangelogFile="true"/>
-	<include file="changelog/data/db.changelog-data-v1.0.0.xml" relativeToChangelogFile="true"/>-->
+	<!--<include file="changelog/data/db.changelog-data-v1.0.0.xml" relativeToChangelogFile="true"/>-->
 	
 </databaseChangeLog>
