--- src/test/java/com/brickncode/backendarchetype/BackendArchetypeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.brickncode.backendarchetype;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendArchetypeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
