+++ src/test/java/com/bytelius/backendarchetype/BackendArchetypeApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bytelius.backendarchetype;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendArchetypeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
