+++ src/test/java/com/bytelius/backendarchetype/controller/impl/SampleControllerTest.java
@@ -1,0 +1,46 @@
+package com.bytelius.backendarchetype.controller.impl;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class SampleControllerTest {
+
+	@BeforeEach
+	void setUp() {
+	}
+
+	@AfterEach
+	void tearDown() {
+	}
+
+	@Test
+	void getAllSampleCityMethod() {
+	}
+
+	@Test
+	void getSampleCityByPathVariableIDMethod() {
+	}
+
+	@Test
+	void getSampleCityByRequestIDParamMethod() {
+	}
+
+	@Test
+	void saveSampleCityMethod() {
+	}
+
+	@Test
+	void updateSampleCityMethod() {
+	}
+
+	@Test
+	void deleteSampleCityByPathVariableIDMethod() {
+	}
+
+	@Test
+	void deleteSampleCityByRequestIDParamMethod() {
+	}
+}
